Added data to the map and moved all calculating to the MainView file
