+++ dashboard/src/components/MainView/MainView.js
@@ -7,13 +7,64 @@ import PollutantDetails from "../Pollutants/PollutantDetails";
 import TopCountries from "../TopCountries/TopCountries";
 import { getLocationData } from "../../api";
 import { WorldMap } from "../Map/WorldMap";
-import { countryList, pollutantIDs, pollutantList } from "../../config";
+import { countryList, pollutantList } from "../../utils/CountryData";
 
 const MainView = () => {
   const navigate = useNavigate();
   const [selectedPollutant, setSelectedPollutant] = useState("");
-  const [mapPollutant, setMapPollutant] = useState("");
   const [searchValue, setSearchValue] = useState("");
+  const [currentPollutant, setCurrentPollutant] = useState(pollutantList[0]);
+  const [currCountryList, setCountryList] = useState(countryList);
+
+  const updateCountry = (country, pollutant, value) => {
+    return { ...country, pollutant: pollutant.name, value: value };
+  };
+
+  const fetchData = async (country, pollutant) => {
+    const response = await getLocationData(country.name, currentPollutant.id);
+    if (response.results) {
+      const pollutantValue = response.results.reduce((acc, result) => {
+        let value;
+        const matchingMeasurement = result.measurements.find(m => m.parameter === pollutant.apiName);
+        if (matchingMeasurement) {
+          value = matchingMeasurement.value;
+        }
+        return Number(value);
+      }, 0);
+      return updateCountry(country, pollutant, pollutantValue)
+    } else {
+      return updateCountry(country, pollutant, 0)
+    }
+  };
+
+  const updatePollutantValues = (pollutant) => {
+    Promise.all(currCountryList.map(country => fetchData(country, pollutant)))
+      .then(newCountryList => setCountryList(newCountryList));
+  };
+
+  useEffect(() => {
+    let timer;
+    if (selectedPollutant) {
+      const newPollutant = pollutantList.find(pollutant => pollutant.name === selectedPollutant)
+      console.log(newPollutant)
+      updatePollutantValues(newPollutant);
+    } else {
+      let counter = 0;
+      timer = setInterval(() => {
+        if (counter >= pollutantList.length) {
+          counter = 0;
+        }
+
+        setCurrentPollutant(pollutantList[counter++]);
+      }, 10000);
+    }
+
+    return () => clearInterval(timer);
+  }, [selectedPollutant]);
+
+  useEffect(() => {
+    updatePollutantValues(currentPollutant);
+  }, [currentPollutant]);
 
   const onSearchButtonClick = async () => {
     /* navigate("/detailed", {
@@ -76,12 +127,12 @@ const MainView = () => {
                 <div className="pollutant-details-container">
                   <PollutantDetails selectedPollutant={selectedPollutant} />
 
-                  <TopCountries selectedPollutant={mapPollutant} />
+                  <TopCountries countryList={currCountryList} />
                 </div>
               </div>
             </div>
             <div className="main-map-container">
-              <WorldMap />
+              <WorldMap countryList={currCountryList} />
             </div>
           </div>
         </div>
